VS2013/64-bit, part 3
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1886 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/Common/pr3287/trace.c
@@ -113,9 +113,9 @@ trace_ds(const char *fmt, ...)
 		nl = true;
 	}
 	while (tscnt + len >= 75) {
-		int plen = 75-tscnt;
+		size_t plen = 75-tscnt;
 
-		(void) fprintf(tracef, "%.*s ...\n... ", plen, s);
+		(void) fprintf(tracef, "%.*s ...\n... ", (int)plen, s);
 		tscnt = 4;
 		s += plen;
 		len -= plen;
